server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["principle105 <principle105@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = {extras = ["email"], version = "^1.9.0"}
beanie = "^1.10.3"
fastapi = "^0.75.0"
motor = "^2.5.1"
Authlib = "^1.0.0"
httpx = "^0.22.0"
fastapi-jwt-auth = "^0.5.0"
uvicorn = "^0.17.6"
python-decouple = "^3.6"
itsdangerous = "^2.1.2"
starlette-csrf = "^1.4.3"
black = "^22.3.0"
isort = "^5.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
